pyproject.toml
[project]
name = "cpa-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apify-client>=2.5.0",
    "docling>=2.88.0",
    "fastapi>=0.135.3",
    "groq>=1.1.2",
    "langgraph>=1.1.6",
    "pydantic>=2.13.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "reportlab>=4.4.10",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.44.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.14.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-mock>=3.15.1",
]
